Update django_filters/filterset.py
By default, django-filter doesn't work well with django-pagination . There is a trick to do this [1], but this breaks the polimorphism of "object_list" (needed  when use the same template sometimes with a  FilterSet and some as a standard queryset). 

[1] https://github.com/alex/django-filter/issues/13

Defining this couple of methods as a fallback to the inner qs, it works out of the box without custom django-pagination syntax.