Fix regression in order_by
Ordering was totally broken! The `ordering_field` was *not* in
`self.filters`, but `ordered_value` was only set while iterating over
`self.filters`. As such, `ordered_value` was always `None`. Because of
this, the default ordering was always used. The tests passed because
none of the tests checked to see if ordering on not the first field in
`Meta.order_by` worked.

The code has been fixed, and ordering is now possible again.

The tests have been fixed, and ordering by all fields in `Meta.order_by`
is now tested.