Fixed error due to double-passing of help_text.
The following error was being raised for non-LookupType filters with
help_text applied.

TypeError: type object got multiple values for keyword argument
'help_text'

The fix is to add the text “This is an exclusion filter” onto the
field’s help_text and put that back in extra carefully. self.extra is
not modified.