Fix the auto-generated documentation for filters.
This makes it so that the documentation generated for query-string
parameters in Django-Rest-Framework generated by django_filters displays
help_text properly.

According to coreapi.document, passing type, description, and example to
coreapi.Field are deprecated and, in fact, the template that DRF uses to
display filters references `field.schema.description` instead of
`field.description`.  To accomodate this, create a coreschema.String
object and pass the description argument to that instead of to the Field
object itself.