docs: Call out strictness changes in 2.0
I'm guessing this wasn't intentional, given that it wasn't called out in
the release notes or original PR. However, 2.0 is out in the wild now so
provide a way for users to revert to the previous behavior, if they so
choose.

All of this should be integrated into a proper document, but that's a
job for another day.

Signed-off-by: Stephen Finucane <stephen@that.guru>
Fixes: 78febd1 ("Rework validation, add queryset filter method (#788)")