Fix detecting crispy_forms from tests
Fix the test logic to correctly detect that crispy_forms are installed.
Currently, this logic relies on the is_crispy() function that returns
True only if crispy_forms are already added to INSTALLED_APPS -- but
we can do that only after checking whether they are available.  Check
on crispy_forms import directly instead.